move SetViewName to separate file
